+++ packages/jest-snapshot/src/__tests__/utils.test.js
@@ -5,11 +5,15 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-jest.mock('fs');
+jest.mock('../bound_fs', () => ({
+  boundExistsSync: jest.fn(() => true),
+  boundReadFile: jest.fn(),
+  boundWriteFile: jest.fn(),
+}));
 
-const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
+const boundFs = require('../bound_fs');
 
 const {
   getSnapshotData,
@@ -23,18 +27,8 @@ const {
   SNAPSHOT_VERSION_WARNING,
 } = require('../utils');
 
-const writeFileSync = fs.writeFileSync;
-const readFileSync = fs.readFileSync;
-const existsSync = fs.existsSync;
 beforeEach(() => {
-  fs.writeFileSync = jest.fn();
-  fs.readFileSync = jest.fn();
-  fs.existsSync = jest.fn(() => true);
-});
-afterEach(() => {
-  fs.writeFileSync = writeFileSync;
-  fs.readFileSync = readFileSync;
-  fs.existsSync = existsSync;
+  jest.clearAllMocks();
 });
 
 test('keyToTestName()', () => {
@@ -57,7 +51,7 @@ test('saveSnapshotFile() works with \r\n', () => {
   };
 
   saveSnapshotFile(data, filename);
-  expect(fs.writeFileSync).toBeCalledWith(
+  expect(boundFs.boundWriteFile).toBeCalledWith(
     filename,
     `// Jest Snapshot v1, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
@@ -71,7 +65,7 @@ test('saveSnapshotFile() works with \r', () => {
   };
 
   saveSnapshotFile(data, filename);
-  expect(fs.writeFileSync).toBeCalledWith(
+  expect(boundFs.boundWriteFile).toBeCalledWith(
     filename,
     `// Jest Snapshot v1, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
@@ -80,7 +74,9 @@ test('saveSnapshotFile() works with \r', () => {
 
 test('getSnapshotData() throws when no snapshot version', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(() => 'exports[`myKey`] = `<div>\n</div>`;\n');
+  boundFs.boundReadFile.mockImplementation(
+    () => 'exports[`myKey`] = `<div>\n</div>`;\n',
+  );
   const update = 'none';
 
   expect(() => getSnapshotData(filename, update)).toThrowError(
@@ -95,7 +91,7 @@ test('getSnapshotData() throws when no snapshot version', () => {
 
 test('getSnapshotData() throws for older snapshot version', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(
+  boundFs.boundReadFile.mockImplementation(
     () =>
       `// Jest Snapshot v0.99, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
@@ -118,7 +114,7 @@ test('getSnapshotData() throws for older snapshot version', () => {
 
 test('getSnapshotData() throws for newer snapshot version', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(
+  boundFs.boundReadFile.mockImplementation(
     () =>
       `// Jest Snapshot v2, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
@@ -141,7 +137,9 @@ test('getSnapshotData() throws for newer snapshot version', () => {
 
 test('getSnapshotData() does not throw for when updating', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(() => 'exports[`myKey`] = `<div>\n</div>`;\n');
+  boundFs.boundReadFile.mockImplementation(
+    () => 'exports[`myKey`] = `<div>\n</div>`;\n',
+  );
   const update = 'all';
 
   expect(() => getSnapshotData(filename, update)).not.toThrow();
@@ -149,7 +147,9 @@ test('getSnapshotData() does not throw for when updating', () => {
 
 test('getSnapshotData() marks invalid snapshot dirty when updating', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(() => 'exports[`myKey`] = `<div>\n</div>`;\n');
+  boundFs.boundReadFile.mockImplementation(
+    () => 'exports[`myKey`] = `<div>\n</div>`;\n',
+  );
   const update = 'all';
 
   expect(getSnapshotData(filename, update)).toMatchObject({dirty: true});
@@ -157,7 +157,7 @@ test('getSnapshotData() marks invalid snapshot dirty when updating', () => {
 
 test('getSnapshotData() marks valid snapshot not dirty when updating', () => {
   const filename = path.join(__dirname, 'old-snapshot.snap');
-  fs.readFileSync = jest.fn(
+  boundFs.boundReadFile.mockImplementation(
     () =>
       `// Jest Snapshot v${SNAPSHOT_VERSION}, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`myKey`] = `<div>\n</div>`;\n',
@@ -171,7 +171,7 @@ test('escaping', () => {
   const filename = path.join(__dirname, 'escaping.snap');
   const data = '"\'\\';
   saveSnapshotFile({key: data}, filename);
-  const writtenData = fs.writeFileSync.mock.calls[0][1];
+  const writtenData = boundFs.boundWriteFile.mock.calls[0][1];
   expect(writtenData).toBe(
     `// Jest Snapshot v1, ${SNAPSHOT_GUIDE_LINK}\n\n` +
       'exports[`key`] = `"\'\\\\`;\n',